Got rid of gateway (Nuked)
We nuked her so that stuff would show up on the front end and got rid of all gateway stuff
