--- IoT-Backend/src/main/java/com/myproject/iot/dto/CreateGatewayPayload.java
@@ -1,15 +1,0 @@
-package com.myproject.iot.dto;
-
-public class CreateGatewayPayload {
-
-    private String name;
-
-    public String getGatewayName() {
-        return name;
-    }
-
-    public void setGatewayName(String name) {
-        this.name = name;
-    }
-}
-
