--- IoT-Backend/src/main/java/com/myproject/iot/service/GatewayService.java
@@ -1,26 +1,0 @@
-package com.myproject.iot.service;
-
-//import all the things
-import com.myproject.iot.domain.Gateway;
-import com.myproject.iot.repository.GatewayRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class GatewayService {
-
-    @Autowired//creates gateway repository instance
-    private GatewayRepository gatewayRepository;
-
-    public Gateway addGateway(String name) {
-        Gateway gateway = new Gateway();
-        gateway.setGatewayName(name);
-        return gatewayRepository.save(gateway);
-    }
-
-    public List<Gateway> getGateways() {
-        return gatewayRepository.findAll();
-    }
-}
